streamlit_wec_full_app_repo/requirements.txt
streamlit
pandas
numpy
matplotlib
openai
gspread
fpdf
google-auth
